reorganize cli commands
