+++ populus/cli/__init__.py
@@ -1,0 +1,6 @@
+from .main import main  # NOQA
+from .chain_cmd import chain  # NOQA
+from .compile_cmd import compile_contracts  # NOQA
+from .deploy_cmd import deploy  # NOQA
+from .init_cmd import init  # NOQA
+from .web_cmd import web  # NOQA
