+++ populus/cli/main.py
@@ -1,0 +1,9 @@
+import click
+
+
+@click.group()
+def main():
+    """
+    Populus
+    """
+    pass
